+++ src/main/java/org/jabref/logic/exporter/MetaDataSerializer.java
@@ -43,6 +43,8 @@ public class MetaDataSerializer {
                 path -> stringyMetaData.put(MetaData.FILE_DIRECTORY, Collections.singletonList(path.trim())));
         metaData.getUserFileDirectories().forEach((user, path) -> stringyMetaData
                 .put(MetaData.FILE_DIRECTORY + '-' + user, Collections.singletonList(path.trim())));
+        metaData.getLaTexFileDirectories().forEach((user, path) -> stringyMetaData
+                .put(MetaData.FILE_DIRECTORY + "Latex-" + user, Collections.singletonList(path.toString().trim())));
 
         for (ContentSelector selector: metaData.getContentSelectorList()) {
                 stringyMetaData.put(MetaData.SELECTOR_META_PREFIX + selector.getFieldName(), selector.getValues());