+++ src/main/java/org/jabref/logic/importer/fileformat/RisImporter.java
@@ -2,6 +2,9 @@ package org.jabref.logic.importer.fileformat;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.time.Year;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -23,6 +26,7 @@ import org.jabref.model.entry.Month;
 public class RisImporter extends Importer {
 
     private static final Pattern RECOGNIZED_FORMAT_PATTERN = Pattern.compile("TY  - .*");
+    private static DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy");
 
     @Override
     public String getName() {
@@ -53,9 +57,11 @@ public class RisImporter extends Importer {
         String linesAsString = reader.lines().reduce((line, nextline) -> line + "\n" + nextline).orElse("");
 
         String[] entries = linesAsString.replace("\u2013", "-").replace("\u2014", "--").replace("\u2015", "--")
-                .split("ER  -.*\\n");
+                                        .split("ER  -.*\\n");
 
         for (String entry1 : entries) {
+
+            boolean foundDate = false;
 
             String type = "";
             String author = "";
@@ -74,7 +80,7 @@ public class RisImporter extends Importer {
                 while (!done && (j < (lines.length - 1))) {
                     if ((lines[j + 1].length() >= 6) && !"  - ".equals(lines[j + 1].substring(2, 6))) {
                         if ((current.length() > 0) && !Character.isWhitespace(current.charAt(current.length() - 1))
-                                && !Character.isWhitespace(lines[j + 1].charAt(0))) {
+                            && !Character.isWhitespace(lines[j + 1].charAt(0))) {
                             current.append(' ');
                         }
                         current.append(lines[j + 1]);
@@ -123,7 +129,7 @@ public class RisImporter extends Importer {
                         fields.put(FieldName.TITLE, fields.get(FieldName.TITLE).replaceAll("\\s+", " ")); // Normalize whitespaces
                     } else if ("BT".equals(tag)) {
                         fields.put(FieldName.BOOKTITLE, value);
-                    } else if (("T2".equals(tag) || "J2".equals(tag) || "JA".equals(tag)) && (fields.get(FieldName.JOURNAL) == null || "".equals(fields.get(FieldName.JOURNAL)))) {
+                    } else if (("T2".equals(tag) || "J2".equals(tag) || "JA".equals(tag)) && ((fields.get(FieldName.JOURNAL) == null) || "".equals(fields.get(FieldName.JOURNAL)))) {
                         //if there is no journal title, then put second title as journal title
                         fields.put(FieldName.JOURNAL, value);
                     } else if ("JO".equals(tag) || "J1".equals(tag) || "JF".equals(tag)) {
@@ -187,8 +193,18 @@ public class RisImporter extends Importer {
                         }
                     } else if ("UR".equals(tag) || "L2".equals(tag) || "LK".equals(tag)) {
                         fields.put(FieldName.URL, value);
-                    } else if (("Y1".equals(tag) || "Y2".equals(tag) || "PY".equals(tag) || "DA".equals(tag)) && (value.length() >= 4)) {
-                        fields.put(FieldName.YEAR, value.substring(0, 4));
+                    } else if (!foundDate && (("Y1".equals(tag) || "Y2".equals(tag) || "PY".equals(tag) || "DA".equals(tag)) && (value.length() >= 4))) {
+                        String year = value.substring(0, 4);
+
+                        try {
+                            Year.parse(year, formatter);
+                            //if the year is parsebale we have found our date
+                            fields.put(FieldName.YEAR, value.substring(0, 4));
+                            foundDate = true;
+                        } catch (DateTimeParseException ex) {
+                            //We can't parse the year, we ignore it
+                        }
+
                         String[] parts = value.split("/");
                         if ((parts.length > 1) && !parts[1].isEmpty()) {
                             try {
@@ -210,7 +226,7 @@ public class RisImporter extends Importer {
                             comment = comment + OS.NEWLINE;
                         }
                         comment = comment + value;
-                    }  else if ("M3".equals(tag) || "DO".equals(tag)) {
+                    } else if ("M3".equals(tag) || "DO".equals(tag)) {
                         addDoi(fields, value);
                     } else if ("C3".equals(tag)) {
                         fields.put(FieldName.EVENTTITLE, value);